Add threadsafe Pool data structure
Also split the now-large ConcurrentUtilities.jl file into more
manageable individual files.
